Update Readme with edge case first time setup
When logging into cloudshell editor for the first time, users are automatically put in the directory ~/cloudshell_open/cloud-shell-tutorials. Updated the instructions to include a `cd` before `git clone` ensuring that users are put in the root directory before starting the tutorial.

Added additional instructions to refresh the page when users run into an issue where the tutorial panel gets stuck in a blank state.